+++ tests/testthat/test-filter.r
@@ -188,7 +188,7 @@ X
 ")
 
   datesDF$X <- as.POSIXlt(datesDF$X)
-  expect_error(filter(datesDF, X > as.POSIXlt("2014-03-13")), "has unsupported type")
+  expect_error(filter(datesDF, X > as.POSIXlt("2014-03-13")), "use `POSIXct` instead")
 })
 
 test_that( "filter handles complex vectors (#436)", {