+++ NEWS.md
@@ -1,5 +1,10 @@
 # dplyr 0.4.3.9000
 
+* `data_frame()` and `as_data_frame()` now check that you don't have any 
+  `POSIXlt` columns, and tell you to use `POSIXct` if you do (#813).
+  `data_frame()` gives slightly better error messages if you have
+  non-1d atomic vectors/lists.
+
 * `chain()`, `chain_q()` and `%.%` have been removed
 
 * `trunc_mat()` correctly prints the type of list columns (#1379)